Multithreading, bugs fix and JUnit tests - JUnit tests for Cubrid, Informix and Ingres - Fix nullable result on MySQL newer version - Database vendor selection is now exclusive - Fix URL decode in Coder manager - Fix opened thread when selecting database - Optimize multithreading of reading file: process partial result - Explicit naming of threads Callable - Fix #4062 - Fix #2943 - Fix #4386 - Code design and SonarQube cleaning

+++ src/main/java/com/jsql/view/swing/bruteforce/Bruter.java
@@ -126,7 +126,7 @@ public class Bruter {
         long timeTaken = calculateTimeDifference();
         seconds = (int) timeTaken;
         if (seconds > 60) {
-            minutes = (int) (seconds / 60);
+            minutes = seconds / 60;
             if (minutes * 60 > seconds) {
                 minutes = minutes - 1;
             }
